--- apps/tutorial/app/config/environment.d.ts
@@ -1,15 +1,0 @@
-export default config;
-
-/**
- * Type declarations for
- *    import config from 'tutorial/config/environment'
- */
-declare const config: {
-  environment: string;
-  modulePrefix: string;
-  podModulePrefix: string;
-  locationType: 'history' | 'hash' | 'none' | 'auto';
-  rootURL: string;
-  APP: Record<string, unknown>;
-  SERVICE_WORKER: boolean;
-};
