--- apps/tutorial/app/config/environment.js
@@ -1,22 +1,0 @@
-import { getGlobalConfig } from '@embroider/macros/src/addon/runtime';
-
-const ENV = {
-  modulePrefix: 'tutorial',
-  environment: import.meta.env.DEV ? 'development' : 'production',
-  rootURL: '/',
-  locationType: 'history',
-  EmberENV: {},
-  APP: {},
-};
-
-export default ENV;
-
-export function enterTestMode() {
-  ENV.locationType = 'none';
-  ENV.APP.rootElement = '#ember-testing';
-  ENV.APP.autoboot = false;
-
-  let config = getGlobalConfig()['@embroider/macros'];
-
-  if (config) config.isTesting = true;
-}
