+++ apps/tutorial/tsconfig.json
@@ -14,7 +14,7 @@
       "tutorial/*": ["./app/*"],
       "*": ["./types/*"]
     },
-    "types": ["ember-source/types", "@embroider/core/virtual"]
+    "types": ["vite/client", "ember-source/types", "@embroider/core/virtual"]
   },
   "include": ["app/**/*", "tests/**/*", "types/**/*"]
 }
